move page logic to container
