+++ frontend/webapp/app/overview/destinations/create/[id]/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-import React from "react";
 import { NewDestinationForm } from "@/containers/overview";
 
 export default function CreateNewDestinationPage() {
