+++ frontend/webapp/app/overview/destinations/create/page.tsx
@@ -1,5 +1,3 @@
-"use client";
-import React from "react";
 import { NewDestinationFlow } from "@/containers/overview";
 
 export default function CreateDestinationPage() {
