+++ frontend/webapp/containers/overview/destination/update.destination.flow.tsx
@@ -1,20 +1,22 @@
 "use client";
-import React, { useMemo } from "react";
+import React, { useEffect, useMemo, useState } from "react";
 import { KeyvalLoader } from "@/design.system";
-import { OVERVIEW, QUERIES } from "@/utils/constants";
+import { NOTIFICATION, OVERVIEW, QUERIES } from "@/utils/constants";
 import { useMutation, useQuery } from "react-query";
 import { getDestination, updateDestination } from "@/services";
 import { ManageDestination } from "@/components/overview";
-import { deleteDestination } from "@/services/destinations";
+import { deleteDestination, getDestinations } from "@/services/destinations";
 import { ManageDestinationWrapper } from "./destination.styled";
-import { useRouter } from "next/navigation";
+import { useRouter, useSearchParams } from "next/navigation";
+import { useNotification } from "@/hooks";
+const DEST = "dest";
 
-export default function UpdateDestinationFlow({
-  selectedDestination,
-  onSuccess,
-  onError,
-}) {
+export function UpdateDestinationFlow({}) {
+  const [selectedDestination, setSelectedDestination] = useState<any>(null);
+
   const router = useRouter();
+  const searchParams = useSearchParams();
+  const { show, Notification } = useNotification();
 
   const manageData = useMemo(() => {
     return {
@@ -22,6 +24,7 @@ export default function UpdateDestinationFlow({
       ...selectedDestination?.destination_type,
     };
   }, [selectedDestination]);
+
   const { isLoading: destinationTypeLoading, data: destinationType } = useQuery(
     [QUERIES.API_DESTINATION_TYPE, selectedDestination?.type],
     () => getDestination(selectedDestination?.type),
@@ -30,6 +33,12 @@ export default function UpdateDestinationFlow({
     }
   );
 
+  const {
+    isLoading: destinationLoading,
+    data: destinationList,
+    refetch,
+  } = useQuery([QUERIES.API_DESTINATIONS], getDestinations);
+
   const { mutate: handleUpdateDestination } = useMutation((body) =>
     updateDestination(body, selectedDestination?.id)
   );
@@ -37,6 +46,8 @@ export default function UpdateDestinationFlow({
   const { mutate: handleDeleteDestination } = useMutation((body) =>
     deleteDestination(selectedDestination?.id)
   );
+
+  useEffect(onPageLoad, [searchParams, destinationList]);
 
   function onDelete() {
     handleDeleteDestination(selectedDestination.id, {
@@ -57,6 +68,36 @@ export default function UpdateDestinationFlow({
     });
   }
 
+  function onPageLoad() {
+    const search = searchParams.get(DEST);
+    const currentDestination = destinationList?.filter(
+      ({ id }) => id === search
+    );
+    if (currentDestination?.length) {
+      setSelectedDestination(currentDestination[0]);
+    }
+  }
+
+  function onSuccess(message = OVERVIEW.DESTINATION_UPDATE_SUCCESS) {
+    refetch();
+    show({
+      type: NOTIFICATION.SUCCESS,
+      message,
+    });
+  }
+
+  function onError({ response }) {
+    const message = response?.data?.message;
+    show({
+      type: NOTIFICATION.ERROR,
+      message,
+    });
+  }
+
+  if (destinationLoading || !selectedDestination) {
+    return <KeyvalLoader />;
+  }
+
   return destinationTypeLoading ? (
     <KeyvalLoader />
   ) : (
@@ -68,6 +109,7 @@ export default function UpdateDestinationFlow({
         onSubmit={onSubmit}
         onDelete={onDelete}
       />
+      <Notification />
     </ManageDestinationWrapper>
   );
 }
