+++ frontend/webapp/containers/overview/index.tsx
@@ -3,3 +3,4 @@ export { DestinationContainer } from "./destination/destination";
 export { SourcesContainer } from "./sources/sources";
 export { NewDestinationFlow } from "./destination/new.destination.flow";
 export { NewDestinationForm } from "./destination/new.destination.form";
+export { UpdateDestinationFlow } from "./destination/update.destination.flow";
